+++ packages/help-center/src/components/help-center-notice.tsx
@@ -3,8 +3,7 @@ import { ExternalLink } from '@wordpress/components';
 import { createInterpolateElement } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { Icon, info } from '@wordpress/icons';
-import { SupportActivity } from '../data/types';
-import type { AnalysisReport } from '../data/types';
+import type { AnalysisReport, SupportActivity } from '../types';
 import type { ReactNode } from 'react';
 
 type Props = {
