+++ packages/help-center/src/data/use-support-availability.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import apiFetch from '@wordpress/api-fetch';
 import wpcomRequest, { canAccessWpcomApis } from 'wpcom-proxy-request';
-import { OtherSupportAvailability, ChatAvailability } from './types';
+import { OtherSupportAvailability, ChatAvailability } from '../types';
 
 type ResponseType< T extends 'CHAT' | 'OTHER' > = T extends 'CHAT'
 	? ChatAvailability
